Added cluster_monolithic
Moved the cutoff for quaternion and principal axis to 30 degrees. Doesn't seem
to change much; helps a bit but not much.

Moved the seed of quaternion and principal axis clustering to be based on SVD
of the entire dataset as if it were one vertex. Doesn't seem to change much.